requirements.txt
aiocron==1.4
asyncio==3.4.3
certifi==2020.12.5
chardet==4.0.0
coinoxr==0.0.1a0
croniter==1.0.8
future==0.18.2
idna==2.10
natsort==7.1.1
pyaes==1.6.1
pyasn1==0.4.8
pycoingecko==1.4.0
pyjokes==0.6.0
pymongo==3.11.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4
requests==2.25.1
rsa==4.7.2
six==1.15.0
Telethon==1.13.0
typing==3.7.4.1
tzlocal==2.1
ujson==5.2.0
urllib3==1.26.5
